third_party/blink/web_tests/fast/events/touch/touch-inside-iframe-scrolled-expected.txt

Test iframes receive touches correctly when scrolled. https://bugs.webkit.org/show_bug.cgi?id=96541

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS correctDiv is true
PASS touch.pageX is 150
PASS touch.pageY is 350
PASS successfullyParsed is true

TEST COMPLETE
